+++ apps/cryptics/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class CrypticsConfig(AppConfig):
-    name = 'cryptics'
+    name = 'apps.cryptics'